Refactor fishing period
