+++ frontend/src/domain/types/regulation.js
@@ -70,9 +70,19 @@
  * @property {boolean} annualRecurrence
  * @property {[DateInterval]} dateRanges
  * @property {[date]} dates
- * @property {[string]} weekDays
+ * @property {[string]} weekdays
  * @property {boolean} holidays
  * @property {TimeInterval} timeIntervals
  * @property {boolean} holidays
  * @property {boolean} daytime
+ * @property {string} otherInfo
+ */
+
+/**
+ * @typedef Species
+ * @property {boolean} authorized
+ * @property {boolean} allSpecies
+ * @property {string} otherInfo
+ * @property {string[]} species - FAO codes
+ * @property {string[]} speciesGroups - group name
  */
