+++ frontend/src/features/backoffice/create_regulation/regulatory_identification/RegulationGeometryLine.js
@@ -1,15 +1,15 @@
 import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import styled, { css } from 'styled-components'
-import { ContentLine } from '../../commonStyles/Backoffice.style'
-import { Label } from '../../commonStyles/Input.style'
-import CustomSelectComponent from './CustomSelectComponent'
-import MenuItem from './MenuItem'
-import Tag from './Tag'
-import { ReactComponent as ShowIconSVG } from '../../icons/oeil_affiche.svg'
-import { ReactComponent as HideIconSVG } from '../../icons/oeil_masque.svg'
-import { COLORS } from '../../../constants/constants'
-import { setSelectedGeometryId } from '../Regulation.slice'
+import { ContentLine } from '../../../commonStyles/Backoffice.style'
+import { Label } from '../../../commonStyles/Input.style'
+import CustomSelectComponent from '../custom_form/CustomSelectComponent'
+import MenuItem from '../custom_form/MenuItem'
+import Tag from '../Tag'
+import { ReactComponent as ShowIconSVG } from '../../../icons/oeil_affiche.svg'
+import { ReactComponent as HideIconSVG } from '../../../icons/oeil_masque.svg'
+import { COLORS } from '../../../../constants/constants'
+import { setSelectedGeometryId } from '../../Regulation.slice'
 
 const RegulationGeometryLine = props => {
   const {
